# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter Ops API",
    "description": "A clean REST API over the Urja Meter Ops portal. Exposes meter details, consumption history, distribution transformers, and the full network hierarchy.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/meters": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "List meters",
        "description": "Returns a paginated list of meters. Supports free-text search across\nmeter ID and serial number. Results are 20 per page.",
        "operationId": "list_meters_api_v1_meters_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search by meter ID or serial number",
              "default": "",
              "title": "Q"
            },
            "description": "Search by meter ID or serial number"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (20 per page)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (20 per page)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of meters with basic attributes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Meters Api V1 Meters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get meter detail",
        "description": "Returns full nameplate data, network hierarchy, and geographic coordinates\nfor a single meter.",
        "operationId": "get_meter_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Meter ID, e.g. J100001",
              "title": "Meter Id"
            },
            "description": "Meter ID, e.g. J100001"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Meter Api V1 Meters  Meter Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/consumption": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get meter consumption history",
        "description": "Returns half-hourly energy readings for the default window (~7 days).\nEach reading includes cumulative kWh, kVAh, and R-phase voltage.\nTimestamps are in DD/MM/YYYY HH:MM format as returned by the portal.",
        "operationId": "get_meter_consumption_api_v1_meters__meter_id__consumption_get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Meter ID, e.g. J100001",
              "title": "Meter Id"
            },
            "description": "Meter ID, e.g. J100001"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Meter Consumption Api V1 Meters  Meter Id  Consumption Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers": {
      "get": {
        "tags": [
          "Transformers"
        ],
        "summary": "List distribution transformers",
        "description": "Returns a paginated list of distribution transformers (DTs) with their\nfeeder assignment and rated capacity.",
        "operationId": "list_transformers_api_v1_transformers_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Transformers Api V1 Transformers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hierarchy": {
      "get": {
        "tags": [
          "Hierarchy"
        ],
        "summary": "Network hierarchy tree",
        "description": "Reconstructs the full Zone \u2192 Circle \u2192 Division \u2192 Subdivision \u2192\nSubstation \u2192 Feeder \u2192 DT hierarchy from the bulk export.\n\nNote: this fetches all 403 meters from the portal on first call\n(single page export). At this dataset size it completes in ~1 s.\nWith millions of meters you'd want a dedicated hierarchy endpoint\nor a cached index.",
        "operationId": "get_hierarchy_api_v1_hierarchy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Hierarchy Api V1 Hierarchy Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Full meter export",
        "description": "Returns the full meter dataset including nameplate, hierarchy, and\ngeo coordinates. Uses the portal's HMAC-signed bulk export endpoint.\nThe portal returns all 403 meters on a single page at the time of writing.",
        "operationId": "export_all_meters_api_v1_export_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Export is paginated; 403 meters fit on page 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Export is paginated; 403 meters fit on page 1"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export All Meters Api V1 Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}